Implement a simple shader library.

+++ cobalt/core/containers/vector.h
@@ -18,13 +18,13 @@ namespace cobalt {
              * @param initialCapacity: The initial capacity of the vector.
              * @return: The created vector.
              */
-            Vector(uint initialCapacity)
+            Vector(uint64 initialCapacity)
                 : heap(HeapAllocator()),
                   data(heap.grab(initialCapacity * sizeof(T))),
-                  indices((uint*) heap.grab(initialCapacity * sizeof(uint))),
+                  indices((uint64*) heap.grab(initialCapacity * sizeof(uint64))),
                   capacity(initialCapacity),
                   size(0),
-                  gaps(Stack<uint>(initialCapacity))
+                  gaps(Stack<uint64>(initialCapacity))
                 {}
             /* Creates a vector, copied from another vector.
              * @param other: The vector to copy.
@@ -33,13 +33,13 @@ namespace cobalt {
             Vector(const Vector<T>& other)
                 : heap(HeapAllocator()),
                   data(heap.grab(other.capacity * sizeof(T))),
-                  indices((uint*) heap.grab(other.capacity * sizeof(uint))),
+                  indices((uint64*) heap.grab(other.capacity * sizeof(uint64))),
                   capacity(other.capacity),
                   size(other.size),
                   gaps(other.gaps)
                 {
                     memcpy(data, other.data, other.size * sizeof(T));
-                    memcpy(indices, other.indices, other.size * sizeof(uint));
+                    memcpy(indices, other.indices, other.size * sizeof(uint64));
                 }
             /* Destroys the vector.
              */
@@ -63,14 +63,14 @@ namespace cobalt {
              * @param index: The index of the element.
              * @return: The element at the given index.
              */
-            T& operator[](uint index) {
+            T& operator[](uint64 index) {
                 return *(T*)((char*) data + indices[index] * sizeof(T));
             }
             /* Returns the element at the given index.
              * @param index: The index of the element.
              * @return: The element at the given index.
              */
-            const T& operator[](uint index) const {
+            const T& operator[](uint64 index) const {
                 return *(T*)((char*) data + indices[index] * sizeof(T));
             }
 
@@ -79,7 +79,7 @@ namespace cobalt {
              */
             void push(const T& element) {
                 if (gaps.getSize() > 0) {
-                    uint index = gaps.pop();
+                    uint64 index = gaps.pop();
                     indices[index] = size;
                     new((T*)((char*) data + index * sizeof(T))) T(element);
                 } else {
@@ -96,7 +96,7 @@ namespace cobalt {
              */
             void push(T&& element) {
                 if (gaps.getSize() > 0) {
-                    uint index = gaps.pop();
+                    uint64 index = gaps.pop();
                     indices[index] = size;
                     new((T*)((char*) data + index * sizeof(T))) T(std::move(element));
                 } else {
@@ -114,7 +114,7 @@ namespace cobalt {
             template <typename... Args>
             void emplace(Args&&... args) {
                 if (gaps.getSize() > 0) {
-                    uint index = gaps.pop();
+                    uint64 index = gaps.pop();
                     indices[index] = size;
                     new((T*)((char*) data + index * sizeof(T))) T(std::forward<Args>(args)...);
                 } else {
@@ -130,14 +130,14 @@ namespace cobalt {
             /* Removes the element at the given index.
              * @param index: The index of the element to remove.
              */
-            void remove(uint index) {
+            void remove(uint64 index) {
                 if (index >= size || size == 0 || index < 0) {
                     throw ContainerException("Index out of bounds");
                 }
                 ((T*)((char*) data + indices[index] * sizeof(T)))->~T();
                 gaps.push(indices[index]);
                 size--;
-                for (uint i = index; i < size; i++) {
+                for (uint64 i = index; i < size; i++) {
                     indices[i] = indices[i + 1];
                 }
             }
@@ -145,24 +145,24 @@ namespace cobalt {
             /* Returns the size of the vector.
              * @return: The size of the vector.
              */
-            uint getSize() const {
+            uint64 getSize() const {
                 return size;
             }
 
             private:
             HeapAllocator heap;     // The heap allocator used to allocate the vector.
             void* data;             // The data of the vector.
-            uint* indices;          // Maps each index to its corresponding block.
-            uint capacity;          // The capacity of the vector.
-            uint size;              // The size of the vector.
-            Stack<uint> gaps;       // The gaps in the vector.
+            uint64* indices;          // Maps each index to its corresponding block.
+            uint64 capacity;          // The capacity of the vector.
+            uint64 size;              // The size of the vector.
+            Stack<uint64> gaps;       // The gaps in the vector.
 
             /* Resizes the vector to the given capacity.
              * @param newCapacity: The new capacity of the vector.
              */
-            void resize(uint newCapacity) {
+            void resize(uint64 newCapacity) {
                 data = heap.resize(data, newCapacity * sizeof(T));
-                indices = (uint*) heap.resize(indices, newCapacity * sizeof(uint));
+                indices = (uint64*) heap.resize(indices, newCapacity * sizeof(uint64));
                 capacity = newCapacity;
             }
         };
